restructure file system database
Previously, the file system based database stored data in a very
hierarchical manner. Pages were stored in subdirectories of users,
widgets in subdirectories of pages etc.

This had some downsides. For example, the page ID would have to be
known in order to access a bookmark. But a bookmark itself doesn't
have a back reference to its page, only to its widget.

The new structure is more flat: pages, widgets and bookmarks live
in sibling directories. They are still stored as subdirectories of
the user they belong to, because the user must alsways be known and
it is not part of our business requirements that data must be able
to be shared directly between users. (At least not on the database
level.)

This new structure also is a little closer to how relational databases
store things in tables. The tables are not organized hierarchically,
but table rows refer to other tables by ID.