app: initialize one page & widget for new users
The initialization is pretty hacky and based on the user not having
any bookmarks. The user may already have a page & a widget even if
there are no bookmarks. However, there is no way to find out about
that without a "get widgets" endpoint. Once that endpoint exists,
this will be cleaned up.